Use float instad of integer in topicSearch
Sometimes a 0.0 float comes as result and breaks the API.

+++ lib/sanbase_web/graphql/schema/types/social_data_types.ex
@@ -68,7 +68,7 @@ defmodule SanbaseWeb.Graphql.SocialDataTypes do
   end
 
   object :chart_data do
-    field(:mentions_count, :integer)
+    field(:mentions_count, :float)
     field(:datetime, non_null(:datetime))
   end
